refactor: Restructure Types and Const Assertions
Shift all label-associated types to label.types.ts, providing a dedicated
location for type definitions. Further, implement 'User' const assertion in
auth.const, improving the consistency and expressiveness of the constants used.
