+++ components/auth/auth.const.ts
@@ -1,0 +1,5 @@
+export type USER = (typeof USER)[keyof typeof USER];
+export const USER = {
+  email: 'email',
+  password: 'password',
+} as const;
