+++ components/label/label.states.ts
@@ -2,11 +2,12 @@ import { DATA_DEMO_LABELS } from '@collections/demo';
 import { IDB_KEY, IDB_STORE } from '@constAssertions/storage';
 import { getDataLabels } from '@lib/queries/queryLabels';
 import { queryEffect } from '@lib/stateLogics/effects/atomEffects/queryEffects';
-import { Labels, Todos } from '@lib/types';
+import { Todos } from '@lib/types';
 import { atomComboBoxQuery, atomFilterSelected } from '@states/comboBoxes';
 import { atomTodoNew } from '@states/todos';
 import { atomUserSession } from '@user/user.states';
 import { atom, atomFamily, selector, selectorFamily } from 'recoil';
+import { Labels } from './label.types';
 
 /**
  * Atom Queries
