+++ lib/queries/queryLabels.ts
@@ -1,5 +1,5 @@
 import { STORAGE_KEY } from '@constAssertions/storage';
-import { Labels } from '@lib/types';
+import { Labels } from '@label/label.types';
 import { fetchWithRetry, queries } from '@stateLogics/utils';
 
 const apiLabels = process.env.NEXT_PUBLIC_API_ENDPOINT_LABELS as string;
