+++ pages/api/v1/labels/[labelId].tsx
@@ -1,11 +1,11 @@
 import { databaseConnect } from '@lib/dataConnections/databaseConnection';
 import Label from '@lib/models/Label';
-import { Labels } from '@lib/types';
 import { NextApiRequest, NextApiResponse } from 'next';
 import { getServerSession } from 'next-auth';
 import { authOptions } from '../auth/[...nextauth]';
 import { OBJECT_ID, RETENTION } from '@constAssertions/data';
 import { sanitize, sanitizeObject, retentionPolicy } from '@stateLogics/utils';
+import { Labels } from '@label/label.types';
 
 const LabelById = async (req: NextApiRequest, res: NextApiResponse) => {
   await databaseConnect();
