+++ pages/api/v1/todos/index.tsx
@@ -2,11 +2,12 @@ import { databaseConnect } from '@lib/dataConnections/databaseConnection';
 import Label from '@lib/models/Label';
 import TodoItem from '@lib/models/Todo/TodoItems';
 import TodoNote from '@lib/models/Todo/TodoNotes';
-import { Labels, Todos } from '@lib/types';
+import { Todos } from '@lib/types';
 import mongoose from 'mongoose';
 import type { NextApiRequest, NextApiResponse } from 'next';
 import { getServerSession } from 'next-auth';
 import { authOptions } from '../auth/[...nextauth]';
+import { Labels } from '@label/label.types';
 
 const Todos = async (req: NextApiRequest, res: NextApiResponse) => {
   await databaseConnect();
